Fix code for refactoring #137
