+++ Examples/StateManagement/redux_tutorial/app/App.js
@@ -1,8 +1,9 @@
 import React from 'react';
-import EntryPoint from './app/EntryPoint';
+import CounterListContainer from './containers/CounterListContainer';
 
 import { createStore } from 'redux';
-import reducers from './app/reducers';
+
+import reducers from './reducers';
 import { Provider } from 'react-redux';
 
 const store = createStore(reducers);
@@ -10,7 +11,7 @@ const store = createStore(reducers);
 const App = () => {
     return (
         <Provider store={store}>
-            <EntryPoint />
+            <CounterListContainer />
         </Provider>
     );
 }
